+++ src/sll/include/sll/cli.h
@@ -8,6 +8,193 @@
  * \group cli
  * \desc Docs!
  */
+/**
+ * \flags subgroup
+ * \name Flags
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ */
+
+
+
+/**
+ * \flags macro var
+ * \name SLL_CLI_FLAG_EXPAND_PATH
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_EXPAND_PATH 1
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_GENERATE_ASSEMBLY
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_GENERATE_ASSEMBLY 2
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_GENERATE_BUNDLE
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_GENERATE_BUNDLE 4
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_GENERATE_COMPILED_OBJECT
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_GENERATE_COMPILED_OBJECT 8
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_HELP
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_HELP 16
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_NO_PATHS
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_NO_PATHS 32
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_NO_RUN
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_NO_RUN 64
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_PRINT_ASSEMBLY
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_PRINT_ASSEMBLY 128
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_PRINT_NODES
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_PRINT_NODES 256
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_STRIP_DEBUG
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_STRIP_DEBUG 512
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_STRIP_NAMES
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_STRIP_NAMES 1024
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_VERBOSE
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_VERBOSE 2048
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_VERSION
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Docs!
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_VERSION 4096
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_RESERVED0
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Internally used as `CLI_FLAG_ASSEMBLY_GENERATED`
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_RESERVED0 8192
+
+
+
+/**
+ * \flags macro var
+ * \name  SLL_CLI_FLAG_RESERVED1
+ * \group cli
+ * \subgroup cli-flag
+ * \desc Internally used as `CLI_FLAG_SINGLE_OUTPUT`
+ * \type __SLL_U32
+ */
+#define SLL_CLI_FLAG_RESERVED1 16384